% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/parent_env_chain.R
\name{parent_env_chain}
\alias{parent_env_chain}
\title{The parent environment chain of a given environment.}
\usage{
parent_env_chain(envir = parent.frame())
}
\arguments{
\item{envir}{environment. By default, the \code{parent.frame()}.}
}
\value{
a list of parent environments, with the first element being
   the immediate parent, and the last element being the empty environment.
}
\description{
The parent environment chain of a given environment.
}
\note{
If a parent environment terminates with an environment who is its
   own parent, this will be the last element instead for the returned list.
}

