% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{pp}
\alias{pp}
\title{Ruby-style string interpolation}
\usage{
pp(..., envir = parent.frame(), sep = "", collapse = "")
}
\arguments{
\item{...}{a character vector, or a (possibly nested) list of character vectors.}

\item{envir}{environment. The environment to fetch values to use in interpolation.}

\item{sep}{character. Passed into the \code{collapse} argument of \code{base::paste}.}

\item{collapse}{character. Passed into the \code{collapse} argument of \code{base::paste}.}
}
\description{
Writing strings for messages with variable content often involves a mess of
nested calls to \code{paste}. Use this function to reduce the clutter created by these.
}
\examples{
\dontrun{
  x <- 5
  pp("we assigned ${x} to x")
  x <- "this time we'll use a string"
  pp("we assigned ${x} to x, and the value of x changed")
  .small_env <- new.env()
  .small_env$x <- "this is inside an environment"
  pp("if we declare an environment, we see that x is ${x} there", envir = .small_env)
}
}

